+++ Source/Entrypoint.cs
@@ -7,34 +7,12 @@ using LCVR.Player;
 
 namespace LCVR;
 
-[LCVRPatch]
-[HarmonyPatch]
-internal class VREntryPoint
-{
-    /// <summary>
-    /// The entrypoint for when you join a game
-    /// </summary>
-    [HarmonyPatch(typeof(StartOfRound), "Start")]
-    [HarmonyPostfix]
-    private static void OnGameEntered()
-    {
-        StartOfRound.Instance.StartCoroutine(Start());
-    }
-
-    private static IEnumerator Start()
-    {
-        Logger.Log("Hello from VR!");
-
-        yield return new WaitUntil(() => StartOfRound.Instance.activeCamera != null);
-    }
-}
-
 [LCVRPatch(LCVRPatchTarget.Universal)]
 [HarmonyPatch]
-internal class UniversalEntryPoint
+internal static class Entrypoint
 {
-    [HarmonyPatch(typeof(StartOfRound), "Start")]
-    [HarmonyPostfix]
+    [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.Start))]
+    [HarmonyPrefix]
     private static void OnGameEntered()
     {
         StartOfRound.Instance.StartCoroutine(Start());
@@ -42,7 +20,7 @@ internal class UniversalEntryPoint
 
     private static IEnumerator Start()
     {
-        Logger.Log("Hello from universal!");
+        Logger.Log("Hello game, I am going to initialize now!");
 
         yield return new WaitUntil(() => StartOfRound.Instance.activeCamera != null);
 
@@ -53,7 +31,7 @@ internal class UniversalEntryPoint
         yield return DNet.Initialize();
     }
 
-    [HarmonyPatch(typeof(StartOfRound), "OnDestroy")]
+    [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.OnDestroy))]
     [HarmonyPostfix]
     private static void OnGameLeave()
     {